# opt: Improve performance of IndexJoin planning