--- a/pkg/sql/opt/memo/memo.go
+++ b/pkg/sql/opt/memo/memo.go
@@ -252,6 +252,12 @@ func (m *Memo) NormExpr(group GroupID) *Expr {
 	return m.groups[group].expr(normExprOrdinal)
 }
 
+// NormOp returns the operator type of NormExpr. See that method's comment for
+// more details.
+func (m *Memo) NormOp(group GroupID) opt.Operator {
+	return m.groups[group].expr(normExprOrdinal).op
+}
+
 // MemoizeNormExpr enters a normalized expression into the memo. This requires
 // the creation of a new memo group with the normalized expression as its first
 // expression. If the expression is already part of an existing memo group, then