--- a/pkg/sql/opt/memo/testdata/logprops/scan
+++ b/pkg/sql/opt/memo/testdata/logprops/scan
@@ -89,6 +89,20 @@ scan a
  ├── prune: (3)
  └── interesting orderings: (+1) (-3)
 
+# Test limited scan with 1 row.
+opt
+SELECT s, x FROM a WHERE x > 1 LIMIT 1
+----
+scan a
+ ├── columns: s:3(string) x:1(int!null)
+ ├── constraint: /1: [/2 - ]
+ ├── limit: 1
+ ├── stats: [rows=1]
+ ├── key: ()
+ ├── fd: ()-->(1,3)
+ ├── prune: (3)
+ └── interesting orderings: (+1) (-3)
+
 # Test case where there are no weak keys available.
 opt
 SELECT d FROM a