--- a/pkg/sql/opt/norm/testdata/rules/combo
+++ b/pkg/sql/opt/norm/testdata/rules/combo
@@ -280,35 +280,11 @@ PruneJoinRightCols
          └── filters [type=bool, outer=(1,6), constraints=(/1: (/NULL - ]; /6: (/NULL - ]), fd=(1)==(6), (6)==(1)]
               └── k = x [type=bool, outer=(1,6), constraints=(/1: (/NULL - ]; /6: (/NULL - ])]
 --------------------------------------------------------------------------------
-GenerateIndexScans (higher cost)
+GenerateIndexScans (no changes)
 --------------------------------------------------------------------------------
-   project
-    ├── columns: s:4(string)
-    └── inner-join
-         ├── columns: k:1(int!null) i:2(int!null) s:4(string) x:6(int!null)
-         ├── key: (6)
-         ├── fd: (1)-->(2,4), (1)==(6), (6)==(1)
-         ├── select
-         │    ├── columns: k:1(int!null) i:2(int!null) s:4(string)
-         │    ├── key: (1)
-         │    ├── fd: (1)-->(2,4)
-  -      │    ├── scan a
-  +      │    ├── index-join a
-         │    │    ├── columns: k:1(int!null) i:2(int) s:4(string)
-         │    │    ├── key: (1)
-  -      │    │    └── fd: (1)-->(2,4)
-  +      │    │    ├── fd: (1)-->(2,4)
-  +      │    │    └── scan a@secondary
-  +      │    │         ├── columns: k:1(int!null) s:4(string)
-  +      │    │         ├── key: (1)
-  +      │    │         └── fd: (1)-->(4)
-         │    └── filters [type=bool, outer=(2), constraints=(/2: (/NULL - ])]
-         │         └── i = (10 - 1) [type=bool, outer=(2), constraints=(/2: (/NULL - ])]
-         ├── scan xy
-         │    ├── columns: x:6(int!null)
-         │    └── key: (6)
-         └── filters [type=bool, outer=(1,6), constraints=(/1: (/NULL - ]; /6: (/NULL - ]), fd=(1)==(6), (6)==(1)]
-              └── k = x [type=bool, outer=(1,6), constraints=(/1: (/NULL - ]; /6: (/NULL - ])]
+--------------------------------------------------------------------------------
+GenerateConstrainedScans (no changes)
+--------------------------------------------------------------------------------
 --------------------------------------------------------------------------------
 GenerateIndexScans (no changes)
 --------------------------------------------------------------------------------
@@ -573,7 +549,7 @@ GenerateIndexScans
               ├── s = 'foo' [type=bool, outer=(4), constraints=(/4: [/'foo' - /'foo']; tight)]
               └── f > 100.0 [type=bool, outer=(3), constraints=(/3: [/100.00000000000001 - ]; tight)]
 ================================================================================
-ConstrainScan
+GenerateConstrainedScans
   Cost: 1.55
 ================================================================================
    project
@@ -826,28 +802,8 @@ EliminateSelect
   + └── filters [type=bool, outer=(2,7), constraints=(/2: (/NULL - ]; /7: (/NULL - ]), fd=(2)==(7), (7)==(2)]
   +      └── y = i [type=bool, outer=(2,7), constraints=(/2: (/NULL - ]; /7: (/NULL - ])]
 --------------------------------------------------------------------------------
-GenerateIndexScans (higher cost)
+GenerateIndexScans (no changes)
 --------------------------------------------------------------------------------
-   semi-join
-    ├── columns: k:1(int!null) i:2(int) f:3(float) s:4(string) j:5(jsonb)
-    ├── key: (1)
-    ├── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
-  - ├── scan a
-  + ├── index-join a
-    │    ├── columns: k:1(int!null) i:2(int) f:3(float) s:4(string) j:5(jsonb)
-    │    ├── key: (1)
-  - │    └── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
-  + │    ├── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
-  + │    └── scan a@secondary
-  + │         ├── columns: k:1(int!null) f:3(float) s:4(string) j:5(jsonb)
-  + │         ├── key: (1)
-  + │         └── fd: (1)-->(3-5), (3,4)~~>(1,5)
-    ├── scan xy
-    │    ├── columns: x:6(int!null) y:7(int)
-    │    ├── key: (6)
-    │    └── fd: (6)-->(7)
-    └── filters [type=bool, outer=(2,7), constraints=(/2: (/NULL - ]; /7: (/NULL - ]), fd=(2)==(7), (7)==(2)]
-         └── y = i [type=bool, outer=(2,7), constraints=(/2: (/NULL - ]; /7: (/NULL - ])]
 --------------------------------------------------------------------------------
 GenerateIndexScans (no changes)
 --------------------------------------------------------------------------------