--- a/pkg/sql/opt/xform/rules/scan.opt
+++ b/pkg/sql/opt/xform/rules/scan.opt
@@ -6,4 +6,4 @@
 # GenerateIndexScans creates alternate Scan expressions for each secondary index
 # on the scanned table.
 [GenerateIndexScans, Explore]
-(Scan $def:* & (CanGenerateIndexScans $def)) => (GenerateIndexScans $def)
+(Scan $def:* & (IsCanonicalScan $def)) => (GenerateIndexScans $def)