Previously, we were generating an IndexJoin+Scan expression for every
secondary index on a table. However, unless the Scan is constrained or
limited, the IndexJoin has almost no chance of being the best plan. And
yet we were paying a sizeable perf penalty for generating the IndexJoins
and then not using them.

The change is to generate IndexJoins "one level up", only when a Scan is
wrapped in a Select or Limit. At this point, they become viable plans that
have a reasonable chance of being selected.

There are significant planning-time improvements from this change. Here are
single CPU numbers showing the effects on optbuild + normalization + exploration
times:

  name                      old time/op  new time/op  delta
  Phases/kv-read/Explore    30.1µs ± 2%  27.7µs ± 3%   -8.12%  (p=0.008 n=5+5)
  Phases/planning1/Explore  15.4µs ± 1%  10.6µs ± 1%  -30.90%  (p=0.008 n=5+5)
  Phases/planning2/Explore  56.8µs ± 2%  31.7µs ± 3%  -44.13%  (p=0.008 n=5+5)
  Phases/planning3/Explore  54.2µs ± 3%  36.7µs ± 3%  -32.21%  (p=0.008 n=5+5)
  Phases/planning4/Explore  67.8µs ± 1%  44.4µs ± 3%  -34.45%  (p=0.008 n=5+5)
  Phases/planning5/Explore  54.0µs ± 3%  35.5µs ± 1%  -34.37%  (p=0.008 n=5+5)
  Phases/planning6/Explore   218µs ± 1%    64µs ± 2%  -70.41%  (p=0.008 n=5+5)