@rytaft 2018-08-21
___
*[pkg/sql/opt/xform/custom_funcs.go, line 52 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/28870#-LKS1RJ6115T6_GAa4fJ:-LKSaBCw3SqjIrmmTL0R:bb7ohn6) ([raw file](https://github.com/cockroachdb/cockroach/blob/246af103fde78e323b0b465de605c173f12d3fda/pkg/sql/opt/xform/custom_funcs.go#L52)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

I used this comment in my tree, maybe this is more clear?
```
// IsCanonicalScan returns true if the given ScanOpDef is an original unaltered
// primary index Scan operator (i.e. unconstrained and not limited).
```
</blockquote></details>

sgtm


<!-- Sent from Reviewable.io -->