@RaduBerinde 2018-08-23
___
*[pkg/sql/opt/xform/index_scan_builder.go, line 130 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/28870#-LKSPpzC7YVu16oFvdiN:-LKbu1l2Aokah2uLNZmh:b-qbxl94) ([raw file](https://github.com/cockroachdb/cockroach/blob/246af103fde78e323b0b465de605c173f12d3fda/pkg/sql/opt/xform/index_scan_builder.go#L130)):*
<details><summary><i>Previously, andy-kimball (Andy Kimball) wrote…</i></summary><blockquote>

OK, I reworked as you suggested. See if it makes it clearer.
</blockquote></details>

Looks great, thanks!


<!-- Sent from Reviewable.io -->